requirements.txt
colored==2.3.0
ffmpeg-python==0.2.0
future==1.0.0
rites==0.4.2
yt-dlp==2025.3.31
